llm.py
import requests

def generate_with_tinyllama(prompt: str, max_length: int = 150):
    try:
        response = requests.post(
            "http://localhost:11434/api/generate",
            json={
                "model": "tinyllama",
                "prompt": prompt,
                "stream": False
            }
        )
        response.raise_for_status()
        result = response.json()
        return result.get("response", "").strip()
    except requests.exceptions.RequestException as e:
        return f"Erreur lors de la génération du résumé : {e}"
